Refactor: Move require_env_vars to service_decorators module
